+++ src/lib/Serv.js
@@ -103,8 +103,8 @@ class BaseRPCMethodHandler {
         if (!this)
             throw new Error('bind of class instance needed');
         const THIZ = this;
-        let method;
         let qstr;
+        let method;
         try {
             qstr = URL.parse(req.url, true).query;
             let compressed = qstr['p'];
@@ -119,6 +119,8 @@ class BaseRPCMethodHandler {
             }
             //invoke the method request
             const ans = await THIZ[method](params);
+            if (!ans.token)
+                throw new Error('no token returned');
             THIZ._ret(res, ans);
         }
         catch (err) {