+++ src/lib/SrvRPC.js
@@ -43,7 +43,7 @@ class HttpRPC {
         if (!params)
             params = {};
         params.method = method;
-        params.token = btoa(this.token);
+        params.token = this.token;
         let str = JSON.stringify(params);
         const compressed = LZString.compressToEncodedURIComponent(str);
         const THIZ = this;
@@ -74,6 +74,7 @@ class HttpRPC {
                     THIZ._log.warn(method + ' ' + str);
                     reject(method + ' ' + str);
                 }
+                THIZ.setToken(resp.token); // saves token
                 resolve(resp.result);
             }) //fetch
                 .catch(function (err) {